+++ components/graph/git-ref.js
@@ -57,8 +57,8 @@ var RefViewModel = function(fullRefName, graph) {
     if (oldNode) oldNode.branchesAndLocalTags.remove(self);
   }, null, "beforeChange");
   this.node.subscribe(function(newNode) {
-    if (newNode) newNode.branchesAndLocalTags.push(self);
-  })
+    if (newNode && newNode.isLocalTag) newNode.branchesAndLocalTags.push(self);
+  });
 };
 module.exports = RefViewModel;
 
@@ -161,7 +161,7 @@ RefViewModel.prototype.createRemoteRef = function(callback) {
       refSpec: this.refName, remoteBranch: this.refName }, function(err) {
         if (!err) {
           var newRef = self.graph.getRef("refs/remotes/" + self.graph.currentRemote() + "/" + self.refName);
-          self.node().branchesAndLocalTags.push(newRef);
+          self.node().remoteTags.push(newRef);
           newRef.node(self.node());
         }
         callback(err);