+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 
 ### Fixed
 - Fix crashes due to submodule parsing [#690](https://github.com/FredrikNoren/ungit/issues/690) [#689](https://github.com/FredrikNoren/ungit/issues/689)
+- Fix duplicate remote tag issues [#686](https://github.com/FredrikNoren/ungit/issues/686)
 
 ### Removed
 - Reverted on hover button effects [#688](https://github.com/FredrikNoren/ungit/issues/688)